Crate odata

Rust oData is library for building OData REST services in Rust. This library is referenced and utilized by the ConnectFour project in the game_server repository for service generation.

What is oData?

OData (Open Data Protocol) is an ISO/IEC approved, OASIS standard that defines a set of best practices for building and consuming RESTful APIs. OData helps you focus on your business logic while building RESTful APIs without having to worry about the various approaches to define request and response headers, status codes, HTTP methods, URL conventions, media types, payload formats, query options, etc. OData also provides guidance for tracking changes, defining functions/actions for reusable procedures, and sending asynchronous/batch requests. [1]

Usage Overview

The library can be used to build REST API based using the Entity Data Model. EntityTypes and EntitySets are declared using macros which provide additional information used to generate metadata and create resource paths. See README for sample usage.

Reexports

pub extern crate serde_derive;

pub extern crate serde_json;

pub extern crate serde;

Modules

edm
This module includes type information for a handful of Edm types. Notably the date types are not yet implemented. Type definition guidelines can be found here

entity
This module includes core functionality needed for declaring EntityTypes and EntitySets

model
Provides the Model class which is used to instantiate EntitySets and their underlying types

service
Provides the wrapper around hyper that is used for creating an OData service.

Macros

defEntity
Declare a new EntityType, generating the metadata for use by the Model and Service class.

defEntitySet
defEntitySet!(name: ident, entity_type: ident) Defines a macro for declaring EntitySets. For instance, defEntitySet!(Dogs, Dog); Declares an EntitySet named "Dogs" that consistes of elements of Entity Dog. CRUD-Q operations are then exposed via implementation of the EntitySet trait for struct.

rust_type
Internal macro. Converts Edm types to the underlying Rust type