Module odata::service

Provides the wrapper around hyper that is used for creating an OData service.

Structs

Service
An instance of an oData service that will serve each model it contains. Constructed using a ServiceBuilder.

ServiceBuilder
Constructs an instance of a Service.

Enums

Error
Error types for handler

Res
Used for routing requests and for simplifying CRUD-Q implementation. Providing the correct HTTP StatusCode is handled by the server.